requirements.txt
glcontext==2.3.3
moderngl==5.6.4
numpy==1.20.3
Pillow==8.2.0
